requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-js-asset==1.2.2
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.2
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
whitenoise==5.3.0
